+++ srcs/can/CANController.cpp
@@ -66,7 +66,7 @@ void	CANController::cleanup() {
 }
 
 // TX handler sending frames in classic CAN format
-void	CANController::sendFrame(uint32_t can_id, 
+void	CANController::sendFrame(uint16_t can_id, 
 			const int8_t* data, uint8_t len) {
 
 	if (!_initialized)
@@ -79,7 +79,7 @@ void	CANController::sendFrame(uint32_t can_id,
 }
 
 // TX handler sending frames in CAN_FD format
-void	CANController::sendFrameFD(uint32_t can_id, 
+void	CANController::sendFrameFD(uint16_t can_id, 
 			const int8_t* data, size_t len) {
 
 	if (!_initialized)
